+++ snips-nlu-ontology-parsers/src/builtin_entity_parser.rs
@@ -1,7 +1,6 @@
 use std::collections::{HashMap, HashSet};
 use std::iter::FromIterator;
 use std::ops::Range;
-use std::time::Instant;
 use std::sync::{Arc, Mutex};
 
 use itertools::Itertools;
@@ -121,46 +120,28 @@ impl BuiltinEntityParser {
         sentence: &str,
         filter_entity_kinds: Option<&[BuiltinEntityKind]>,
     ) -> Vec<BuiltinEntity> {
-        lazy_static! {
-            static ref CACHED_ENTITY: Mutex<EntityCache> = Mutex::new(EntityCache::new(60));
-        }
-        let key = CacheKey {
-            lang: self.lang.to_string(),
-            input: sentence.into(),
-            kinds: filter_entity_kinds
-                .map(|kinds|
-                    self.supported_entity_kinds.clone()
-                        .into_iter()
-                        .filter(|entity_kind| kinds.contains(&entity_kind))
-                        .collect())
-                .unwrap_or_else(|| self.supported_entity_kinds.clone()),
-        };
-        CACHED_ENTITY
-            .lock()
-            .unwrap()
-            .cache(&key, |key| {
-                let context = ResolverContext::default();
-                let kind_order = key.kinds.iter()
-                    .map(|kind| kind.into_builtin())
-                    .collect::<Vec<OutputKind>>();
-                self.parser
-                    .parse_with_kind_order(&sentence.to_lowercase(), &context, &kind_order)
-                    .unwrap_or_else(|_| vec![])
-                    .iter()
-                    .filter_map(|m| {
-                        let entity_kind = BuiltinEntityKind::from_rustling(&m.value);
-                        key.kinds.iter().find(|kind| **kind == entity_kind).map(|kind| {
-                            BuiltinEntity {
-                                value: sentence[m.byte_range.0..m.byte_range.1].into(),
-                                range: m.char_range.0..m.char_range.1,
-                                entity: m.value.clone().into_builtin(),
-                                entity_kind: kind.clone(),
-                            }
-                        })
-                    })
-                    .sorted_by(|a, b| Ord::cmp(&a.range.start, &b.range.start))
-            })
-            .entities
+        let context = ResolverContext::default();
+        let kind_order = self.supported_entity_kinds
+            .iter()
+            .filter(|entity_kind|
+                filter_entity_kinds
+                    .map(|kinds| kinds.contains(&entity_kind))
+                    .unwrap_or(true))
+            .map(|kind| kind.into_builtin())
+            .collect::<Vec<OutputKind>>();
+
+        self.parser
+            .parse_with_kind_order(&sentence.to_lowercase(), &context, &kind_order)
+            .unwrap_or_else(|_| vec![])
+            .iter()
+            .map(|m|
+                BuiltinEntity {
+                    value: sentence[m.byte_range.0..m.byte_range.1].into(),
+                    range: m.char_range.0..m.char_range.1,
+                    entity: m.value.clone().into_builtin(),
+                    entity_kind: BuiltinEntityKind::from_rustling(&m.value),
+                })
+            .sorted_by(|a, b| Ord::cmp(&a.range.start, &b.range.start))
     }
 }
 
@@ -189,62 +170,6 @@ fn get_ranges_mapping(tokens_ranges: &Vec<Range<usize>>) -> HashMap<usize, usize
             },
         ));
     ranges_mapping
-}
-
-struct EntityCache {
-    container: HashMap<CacheKey, CacheValue>,
-    valid_duration_sec: u64,
-}
-
-impl EntityCache {
-    fn new(valid_duration_sec: u64) -> EntityCache {
-        EntityCache {
-            container: HashMap::new(),
-            valid_duration_sec,
-        }
-    }
-
-    fn cache<F: Fn(&CacheKey) -> Vec<BuiltinEntity>>(
-        &mut self,
-        key: &CacheKey,
-        producer: F,
-    ) -> CacheValue {
-        let cached_value = self.container.get(key).map(|a| a.clone());
-        if let Some(value) = cached_value {
-            if value.is_valid(self.valid_duration_sec) {
-                return value;
-            }
-        }
-        let value = CacheValue::new(producer(key));
-        self.container.insert(key.clone(), value.clone());
-        value
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct CacheKey {
-    lang: String,
-    input: String,
-    kinds: Vec<BuiltinEntityKind>,
-}
-
-#[derive(Debug, Clone)]
-struct CacheValue {
-    entities: Vec<BuiltinEntity>,
-    instant: Instant,
-}
-
-impl CacheValue {
-    fn new(entities: Vec<BuiltinEntity>) -> CacheValue {
-        CacheValue {
-            entities,
-            instant: Instant::now(),
-        }
-    }
-
-    fn is_valid(&self, valid_duration_sec: u64) -> bool {
-        self.instant.elapsed().as_secs() < valid_duration_sec
-    }
 }
 
 #[cfg(test)]
@@ -337,44 +262,6 @@ mod test {
                 "二 千 十三 年二 月十 日の カリフォルニア州の天気予報は？",
                 None,
             )
-        );
-    }
-
-    #[test]
-    fn test_entity_cache() {
-        fn parse(_: &CacheKey) -> Vec<BuiltinEntity> {
-            vec![
-                BuiltinEntity {
-                    value: "two".into(),
-                    range: 23..26,
-                    entity_kind: BuiltinEntityKind::Number,
-                    entity: SlotValue::Number(NumberValue { value: 2.0 }),
-                },
-                BuiltinEntity {
-                    value: "4.5".into(),
-                    range: 34..42,
-                    entity_kind: BuiltinEntityKind::Number,
-                    entity: SlotValue::Number(NumberValue { value: 4.5 }),
-                },
-            ]
-        }
-
-        let key = CacheKey {
-            lang: "en".into(),
-            input: "test".into(),
-            kinds: vec![],
-        };
-
-        let mut cache = EntityCache::new(10); // caching for 10s
-        assert_eq!(
-            cache.cache(&key, parse).instant,
-            cache.cache(&key, parse).instant
-        );
-
-        let mut cache = EntityCache::new(0); // no caching
-        assert_ne!(
-            cache.cache(&key, parse).instant,
-            cache.cache(&key, parse).instant
         );
     }