+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.56.1] - 2018-06-05
+
+### Added
+- Builtin entity ontology export API
+
+### Changed
+- Bump `snips-nlu-utils` to `0.6.1`
+
+### Removed
+- Builtin entity caching
+
 ## [0.56.0] - 2018-05-09
 
 ### Fixed
@@ -37,6 +48,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Updated Rustling ontology to `0.16.4`
 
+[0.56.1]: https://github.com/snipsco/snips-nlu-ontology/compare/0.56.0...0.56.1
 [0.56.0]: https://github.com/snipsco/snips-nlu-ontology/compare/0.55.0...0.56.0
 [0.55.0]: https://github.com/snipsco/snips-nlu-ontology/compare/0.54.3...0.55.0
 [0.54.3]: https://github.com/snipsco/snips-nlu-ontology/compare/0.54.2...0.54.3